requirements.txt
fastapi
uvicorn
google-cloud-vision
python-multipart
python-jose[cryptography]
passlib
